Builds site on push using github actions
This commit makes use of github actions to build the site on the
gituhb provided CI/CD environment.

It deletes the documentation from version control so that we can only
keep the sources there - cleaner history.

It also deletes the deploy script and moves the logic within the `npm
run build` command in `package.json`